ajout page drink
