+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import './App.scss'
 import 'bootstrap/dist/css/bootstrap.min.css'
 
 import { Outlet } from "react-router-dom"
-import Header from "./components/Header"
+import Header from "./components/header/Header"
 
 function App() {
   return (
