+++ frontend/src/components/header/Header.scss
@@ -1,0 +1,9 @@
+.links {
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+}
+
+.active{
+    text-decoration: underline;
+}
