+++ frontend/src/pages/drinks/Drinks.tsx
@@ -1,0 +1,3 @@
+export default function Drinks() {
+    return (<h1>Drinks page</h1>)
+}
