+++ frontend/src/routes.tsx
@@ -2,6 +2,7 @@ import { createBrowserRouter } from "react-router-dom";
 import App from "./App";
 import Home from "./pages/home/Home";
 import Login from "./pages/Login";
+import Drinks from "./pages/drinks/Drinks";
 
 const router = createBrowserRouter([
   {
@@ -12,6 +13,10 @@ const router = createBrowserRouter([
         path: '',
         element: <Home />,
       },
+      {
+        path: 'drinks',
+        element: <Drinks />
+      }
     ]
   },
   {
